refactor: changed client file directories and fixed style of export icon
