+++ frontend/src/components/clients/dashboard/useProfileDashboard.ts
@@ -1,7 +1,7 @@
 "use client";
 
 import { useState } from "react";
-import { type ClientDoc } from "@/components/clients/ClientList";
+import { type ClientDoc } from "@/components/clients/list/ClientList";
 import { restoreClient, archiveClient } from "@/application/ClientManagementService";
 import { toast } from "sonner";
 import { type TabId } from "./ClientProfileDashboard";
