+++ frontend/src/components/clients/list/ClientFilterBar.tsx
