+++ frontend/src/components/clients/list/FilterableHeaderCell.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { type ClientDoc } from "@/components/clients/ClientList";
+import { type ClientDoc } from "@/components/clients/list/ClientList";
 import { IconFunnel, IconSortAsc, IconSortDesc } from "@/components/ui/Icons";
 
 interface FilterableHeaderCellProps {
