+++ frontend/src/components/clients/tabs/FinancialAidTab.tsx
@@ -14,7 +14,7 @@ import {
   type FinancialAidApplication,
   type PaymentInstallment,
 } from "@/schemas/clientSchema";
-import type { ClientDoc } from "@/components/clients/ClientList";
+import type { ClientDoc } from "@/components/clients/list/ClientList";
 
 // ─── Props ────────────────────────────────────────────────────────────────────
