+++ frontend/src/components/clients/tabs/MedicalTab.tsx
@@ -16,7 +16,7 @@ import {
   type Vaccination,
   type Hospitalization,
 } from "@/schemas/clientSchema";
-import type { ClientDoc } from "@/components/clients/ClientList";
+import type { ClientDoc } from "@/components/clients/list/ClientList";
 
 // ─── Props ────────────────────────────────────────────────────────────────────
