+++ frontend/src/components/clients/tabs/ProfileTab.tsx
@@ -15,7 +15,7 @@ import {
   type BasicInfoFormData,
   type Dependent,
 } from "@/schemas/clientSchema";
-import type { ClientDoc } from "@/components/clients/ClientList";
+import type { ClientDoc } from "@/components/clients/list/ClientList";
 
 // ─── Props ────────────────────────────────────────────────────────────────────
