+++ frontend/src/components/clients/tabs/QuestionnaireTab.tsx
@@ -10,7 +10,7 @@ import { db } from "@/firebase/firebase";
 import { z } from "zod";
 import { AccordionSection } from "@/components/ui/AccordionSection";
 import { questionnaireSchema } from "@/schemas/clientSchema";
-import type { ClientDoc } from "@/components/clients/ClientList";
+import type { ClientDoc } from "@/components/clients/list/ClientList";
 
 // ─── Props ────────────────────────────────────────────────────────────────────
