+++ frontend/src/components/ui/Icons.tsx
@@ -116,9 +116,17 @@ export function IconSortDesc() {
 // ─── Export icon ─────────────────────────────────────────────────────────────
 export function IconExport({ className = "h-4 w-4" }: { className?: string }) {
   return (
-    <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20" fill="currentColor" className={className} aria-hidden="true">
-      <path d="M10.75 2.75a.75.75 0 0 0-1.5 0v8.614L6.295 8.235a.75.75 0 1 0-1.09 1.03l4.25 4.5a.75.75 0 0 0 1.09 0l4.25-4.5a.75.75 0 0 0-1.09-1.03l-2.955 3.129V2.75Z" />
-      <path d="M3.5 12.75a.75.75 0 0 0-1.5 0v2.5A2.75 2.75 0 0 0 4.75 18h10.5A2.75 2.75 0 0 0 18 15.25v-2.5a.75.75 0 0 0-1.5 0v-2.5c0 .69-.56 1.25-1.25 1.25H4.75c-.69 0-1.25-.56-1.25-1.25v-2.5Z" />
+    <svg 
+      xmlns="http://www.w3.org/2000/svg" 
+      viewBox="0 0 20 20" 
+      fill="currentColor" 
+      className={className} 
+      aria-hidden="true"
+    >
+      {/* The container box with an open top edge */}
+      <path d="M4.75 8A2.75 2.75 0 0 0 2 10.75v4.5A2.75 2.75 0 0 0 4.75 18h10.5A2.75 2.75 0 0 0 18 15.25v-4.5A2.75 2.75 0 0 0 15.25 8h-1.5a.75.75 0 0 0 0 1.5h1.5c.69 0 1.25.56 1.25 1.25v4.5c0 .69-.56 1.25-1.25 1.25H4.75c-.69 0-1.25-.56-1.25-1.25v-4.5c0-.69.56-1.25 1.25-1.25h1.5a.75.75 0 0 0 0-1.5h-1.5Z" />
+      {/* The clean arrow pointing straight UP out of the box */}
+      <path d="M10 2.25a.75.75 0 0 1 .53.22l2.5 2.5a.75.75 0 1 1-1.06 1.06L10.75 4.81v6.44a.75.75 0 0 1-1.5 0V4.81L8.03 6.03a.75.75 0 0 1-1.06-1.06l2.5-2.5a.75.75 0 0 1 .53-.22Z" />
     </svg>
   );
 }
